Refactor: Simplify methods on cli and profile
The methods in `cli` and `profile` were doing too much. Some of the
methods have been refactored to be composed of smaller more specific
methods. There is still more refactoring to be done, for example the
`UI.ask` method could take a block and yield on success. This would
simplify the `make_directory` method of `cli`